chore: split chunkMesh to ChunkMeshImpl
pr: https://github.com/MovingBlocks/Terasology/pull/4786
