+++ engine/src/main/java/org/terasology/engine/rendering/logic/ChunkMeshRenderer.java
@@ -3,7 +3,8 @@
 
 package org.terasology.engine.rendering.logic;
 
-import org.lwjgl.opengl.GL15;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.terasology.engine.entitySystem.entity.EntityRef;
 import org.terasology.engine.entitySystem.entity.lifecycleEvents.BeforeDeactivateComponent;
 import org.terasology.engine.entitySystem.entity.lifecycleEvents.OnActivatedComponent;
@@ -19,6 +20,7 @@ import org.terasology.engine.rendering.world.RenderableWorld;
 import org.terasology.engine.world.chunks.RenderableChunk;
 
 import java.lang.ref.PhantomReference;
+import java.lang.ref.Reference;
 import java.lang.ref.ReferenceQueue;
 import java.util.Collection;
 import java.util.HashMap;
@@ -33,12 +35,14 @@ import java.util.Set;
  */
 @RegisterSystem(RegisterMode.CLIENT)
 public class ChunkMeshRenderer extends BaseComponentSystem implements UpdateSubscriberSystem {
+    private static final Logger logger = LoggerFactory.getLogger(ChunkMeshRenderer.class);
+
     @In
     private RenderableWorld renderableWorld;
 
     private Map<EntityRef, EntityBasedRenderableChunk> pseudoChunks = new HashMap<>();
     private ReferenceQueue<ChunkMesh> disposalQueue = new ReferenceQueue<>();
-    private Set<BuffersReference> disposalSet = new HashSet<>();
+    private Set<DisposableChunkMesh> disposalSet = new HashSet<>();
 
     @Override
     public void initialise() {
@@ -48,7 +52,7 @@ public class ChunkMeshRenderer extends BaseComponentSystem implements UpdateSubs
     @ReceiveEvent(components = {ChunkMeshComponent.class, LocationComponent.class})
     public void onNewMesh(OnActivatedComponent event, EntityRef entity, ChunkMeshComponent mesh) {
         pseudoChunks.put(entity, new EntityBasedRenderableChunk(entity));
-        disposalSet.add(new BuffersReference(mesh.mesh, disposalQueue));
+        disposalSet.add(new DisposableChunkMesh(mesh.mesh, disposalQueue));
     }
 
     @ReceiveEvent(components = {MeshComponent.class, LocationComponent.class})
@@ -62,29 +66,31 @@ public class ChunkMeshRenderer extends BaseComponentSystem implements UpdateSubs
 
     @Override
     public void update(float delta) {
-        BuffersReference reference = (BuffersReference) disposalQueue.poll();
-        while (reference != null) {
-            GL15.glDeleteBuffers(reference.bufferIds);
-            disposalSet.remove(reference);
-            reference = (BuffersReference) disposalQueue.poll();
+        Reference<? extends ChunkMesh> reference = null;
+        while ((reference = disposalQueue.poll()) != null) {
+            DisposableChunkMesh hook = (DisposableChunkMesh) reference;
+            hook.cleanup();
         }
     }
 
     @Override
     public void shutdown() {
-        for (BuffersReference reference : disposalSet) {
-            GL15.glDeleteBuffers(reference.bufferIds);
+        for (DisposableChunkMesh reference : disposalSet) {
+            reference.cleanup();
         }
     }
 
-    private static class BuffersReference extends PhantomReference<ChunkMesh> {
-        int[] bufferIds;
+    private class DisposableChunkMesh extends PhantomReference<ChunkMesh> {
+        private final ChunkMesh.DisposableHook hook;
 
-        BuffersReference(ChunkMesh mesh, ReferenceQueue<ChunkMesh> queue) {
-            super(mesh, queue);
-            bufferIds = new int[mesh.vertexBuffers.length + mesh.idxBuffers.length];
-            System.arraycopy(mesh.vertexBuffers, 0, bufferIds, 0, mesh.vertexBuffers.length);
-            System.arraycopy(mesh.idxBuffers, 0, bufferIds, mesh.vertexBuffers.length, mesh.idxBuffers.length);
+        public DisposableChunkMesh(ChunkMesh referent, ReferenceQueue<? super ChunkMesh> q) {
+            super(referent, q);
+            hook = referent.disposalHook();
+        }
+
+        public void cleanup() {
+            hook.dispose();
         }
     }
+
 }
